requirements.txt
requests
beautifulsoup4
pandas
numpy
yfinance
Pillow
matplotlib
python-telegram-bot
sqlalchemy
flask
gunicorn
